requirements.txt
SpeechRecognition
pydub
pocketsphinx
newspaper3k
gTTS
nltk
PyAudio
[html_clean]
